+++ c-ares-resolver/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- `queue_wait_empty()` now takes `Option<Duration>`
 - Return errors instead of panicking when string arguments contain null bytes.
   The following methods now return `Result`s:
   - `Options::set_domains`